moving functions used both in train and eval to a more general utils.py
